New: expose buildDocs() for programmatic documentation builds
